# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "hiai-docs API",
    "description": "Self-hosted AI-first documentation platform. Full-text + semantic search, version history, sharing, and folder organization.",
    "version": "0.2.3",
    "contact": {
      "name": "hiai-gg",
      "url": "https://github.com/hiai-gg/hiai-docs"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "paths": {
...
    "/api/share/{token}": {
      "get": {
        "operationId": "getApiShareByToken",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "token",
            "required": true
          }
        ],
        "responses": {
          "200": {}
        }
      }
    },
    "/api/share/{id}": {
      "delete": {
        "operationId": "deleteApiShareById",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {}
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}